Extract `nativeCompileOptions` to `llvm` package
They're now defined under the name `defaultCompileOptions`.
